try keeping the meta panel on top, outside timeline
